+++ crates/solver/src/settlement_submission/dry_run.rs
@@ -1,7 +1,9 @@
 //! Dry run settlement submission strategy. I.e. just log!
 
-use super::retry::settle_method_builder;
-use crate::{settlement::Settlement, settlement_simulation::tenderly_link};
+use crate::{
+    settlement::Settlement, settlement_simulation::settle_method_builder,
+    settlement_simulation::tenderly_link,
+};
 use anyhow::Result;
 use contracts::GPv2Settlement;
 use ethcontract::Account;
